requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyrogram==2.0.106
tgcrypto==1.2.5
python-dotenv==1.0.0
